fix/Coding standard (#467)

+++ src/Application/Console/Formatters/Console.php
@@ -271,7 +271,7 @@ final class Console implements Formatter
         $lines = [];
         foreach (self::CODE_METRIC_CLASSES as $metric) {
             $name = explode('\\', $metric);
-            $lines[end($name)] = (float) (new $metric())->getPercentage($insightCollection->getCollector());
+            $lines[end($name)] = (new $metric())->getPercentage($insightCollection->getCollector());
         }
 
         $this->writePercentageLines($lines);
@@ -313,7 +313,7 @@ final class Console implements Formatter
         $lines = [];
         foreach (self::ARCHITECTURE_METRIC_CLASSES as $metric) {
             $name = explode('\\', $metric);
-            $lines[end($name)] = (float) (new $metric())->getPercentage($insightCollection->getCollector());
+            $lines[end($name)] = (new $metric())->getPercentage($insightCollection->getCollector());
         }
 
         $this->writePercentageLines($lines);
